feat: seperating a useCryptKeeper() react hook
